[YS-473] refactor: 연구자 내정보수정 연락 받을 이메일 중복 확인 로직 개선
